Suppression d'un niveau d'abstraction pour les fonctions helpers et plutôt mise en snippet
