+++ src/components/testHelper.tsx
@@ -1,6 +1,5 @@
-import { fireEvent, render, RenderResult, screen, within } from '@testing-library/react'
+import { render, RenderResult } from '@testing-library/react'
 import { ReactElement } from 'react'
-import { select } from 'react-select-event'
 import { ToastContainer } from 'react-toastify'
 import { Mock } from 'vitest'
 
@@ -127,34 +126,4 @@ export function stubbedConceal() {
 
 export function stubbedServerAction(result: ReadonlyArray<string>): Mock<() => Promise<ReadonlyArray<string>>> {
   return vi.fn<() => Promise<ReadonlyArray<string>>>().mockResolvedValueOnce(result)
-}
-
-export function presserLeBouton(name: string, description?: string): HTMLElement {
-  const button = screen.getByRole('button', { description, name })
-  fireEvent.click(button)
-  return button
-}
-
-export function presserLeBoutonDans(context: HTMLElement, name: string): HTMLElement {
-  const button = within(context).getByRole('button', { name })
-  fireEvent.click(button)
-  return button
-}
-
-export function cocherLaCase(name: string): void {
-  fireEvent.click(screen.getByRole('checkbox', { name }))
-}
-
-export function presserLeBoutonRadio(name: string): void {
-  fireEvent.click(screen.getByRole('radio', { name }))
-}
-
-export function saisirLeTexte(name: string | RegExp, value: string): HTMLElement {
-  const input = screen.getByLabelText(name)
-  fireEvent.change(input, { target: { value } })
-  return input
-}
-
-export async function selectionnerLElement(input: HTMLElement, nomStructure: string): Promise<void> {
-  await select(input, nomStructure)
 }
